requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.228
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
matplotlib==3.7.3
pyyaml==6.0.1
requests==2.31.0
scipy==1.11.3
